requirements.txt
spotipy
pillow
requests